refact: section definition component + i18n
