+++ src/js/components/Provenance/Modal/SideNav.tsx
@@ -1,27 +1,31 @@
-import type { NavEntry, SectionId } from './types';
+import { useTranslationFn } from '@/hooks';
+import type { ProvenanceEntry, SectionId } from './types';
 
 type SideNavProps = {
-  navEntries: NavEntry[];
+  navEntries: ProvenanceEntry[];
   activeSection: SectionId;
   onJump: (id: SectionId) => void;
 };
 
-const SideNav = ({ navEntries, activeSection, onJump }: SideNavProps) => (
-  <nav className="pm-nav">
-    <div className="pm-nav-cap">On this page</div>
-    {navEntries.map(({ id, label, icon, count }) => (
-      <button
-        key={id}
-        type="button"
-        className={`pm-nav-link${activeSection === id ? ' active' : ''}`}
-        onClick={() => onJump(id)}
-      >
-        {icon}
-        {label}
-        {count !== undefined && <span className="pm-nav-count">{count}</span>}
-      </button>
-    ))}
-  </nav>
-);
+const SideNav = ({ navEntries, activeSection, onJump }: SideNavProps) => {
+  const t = useTranslationFn();
+  return (
+    <nav className="pm-nav">
+      <div className="pm-nav-cap">On this page</div>
+      {navEntries.map(({ id, icon, count }) => (
+        <button
+          key={id}
+          type="button"
+          className={`pm-nav-link${activeSection === id ? ' active' : ''}`}
+          onClick={() => onJump(id)}
+        >
+          {icon}
+          {t(`provenance.sections.${id}`)}
+          {count !== undefined && <span className="pm-nav-count">{count}</span>}
+        </button>
+      ))}
+    </nav>
+  );
+}
 
 export default SideNav;
