+++ src/js/components/Provenance/Modal/cards/index.ts
@@ -1,6 +1,8 @@
+import ProvenanceSection from './ProvenanceSection';
+
 export { CopyButton } from './CopyButton';
 export { OntologyChip } from './OntologyChip';
-export { SectionHead } from './SectionHead';
+export { ProvenanceSection };
 export { LinkTile } from './LinkTile';
 export { PersonCard } from './PersonCard';
 export { PublicationCard } from './PublicationCard';
